--- components/Header/styles.js
@@ -1,5 +1,0 @@
-import Heading from "../Heading";
-
-export default function Header() {
-  return <Heading></Heading>;
-}
