+++ components/Icons/AdminIcon/index.js
