+++ components/Icons/AdminIcon/styles.js
