+++ components/Icons/RoomIcon/styles.js
