+++ components/Icons/StudentIcon/index.js
