+++ components/Icons/StudentIcon/styles.js
