+++ components/Icons/TeacherIcon/index.js
