+++ components/Icons/TeacherIcon/styles.js
